+++ components/org.wso2.carbon.identity.idp.metadata.saml2/src/main/java/org/wso2/carbon/identity/idp/metadata/saml2/util/BuilderUtil.java
@@ -23,7 +23,7 @@ import org.apache.commons.logging.LogFactory;
 import org.opensaml.core.xml.config.XMLObjectProviderRegistrySupport;
 import org.opensaml.core.config.InitializationException;
 import org.opensaml.core.xml.XMLObjectBuilderFactory;
-import org.wso2.carbon.identity.core.util.SAMLInitializer;
+import org.wso2.carbon.identity.saml.common.util.SAMLInitializer;
 import org.wso2.carbon.idp.mgt.MetadataException;
 import javax.xml.namespace.QName;
